server/requirements.txt
fastapi
uvicorn[standard]
pydantic
PyJWT
mysql-connector-python
tensorflow-cpu
h5py
Pillow
numpy
python-multipart
python-dotenv
gunicorn
uvicorn
